refactor(eval): extract context_path.rs
